refactor(luascript): move lua meta to own file
