+++ src/actions.cpp
@@ -9,6 +9,7 @@
 #include "configmanager.h"
 #include "game.h"
 #include "housetile.h"
+#include "lua/meta.h"
 #include "spells.h"
 
 extern Game g_game;
