+++ vc18/atlas.vcxproj
@@ -163,6 +163,7 @@
     <ClCompile Include="..\src\items.cpp" />
     <ClCompile Include="..\src\luascript.cpp" />
     <ClCompile Include="..\src\lua\error.cpp" />
+    <ClCompile Include="..\src\lua\meta.cpp" />
     <ClCompile Include="..\src\mailbox.cpp" />
     <ClCompile Include="..\src\main.cpp" />
     <ClCompile Include="..\src\map.cpp" />
@@ -258,6 +259,7 @@
     <ClInclude Include="..\src\luascript.h" />
     <ClInclude Include="..\src\luavariant.h" />
     <ClInclude Include="..\src\lua\error.h" />
+    <ClInclude Include="..\src\lua\meta.h" />
     <ClInclude Include="..\src\mailbox.h" />
     <ClInclude Include="..\src\map.h" />
     <ClInclude Include="..\src\matrixarea.h" />
